python_service/requirements.txt
paho-mqtt
influxdb-client
websockets
# asyncio ya viene incluido en Python, no es necesario ponerlo aquí
